adding swagger support v2

+++ app/Http/Controllers/Auth/AuthenticatedSessionController.php
@@ -65,7 +65,11 @@ class AuthenticatedSessionController extends Controller
      */
     public function destroy(Request $request): JsonResponse
     {
-        $request->user()->currentAccessToken()->delete();
+        if(method_exists($request->user()->currentAccessToken(), 'delete')) {
+            $request->user()->currentAccessToken()->delete();
+        } else {
+            auth()->guard('web')->logout();
+        }
 
         return response()->json(['message' => 'Logged out']);
     }
